[AMD][BACKEND] Lower predicated load/store to func call that gets lowered at the end to LLVM control flow. (#3767)
The main problem here is that predicated loads/stores used to add
branches, which increases the number of basic blocks in the IR. On some
kernels, the number of basic blocks becomes large, and in specific
scenarios where these basic blocks are mergeable, the MLIR canonicalizer
get stuck merging those blocks. This PR is a temporary workaround. The
long term solution might involve different types of pointers where we
can use MUBUF instructions that have OOB checks, so we won't have to
handle branches. This PR does:

1. Lower Loads/Stores (both shared and global) to LLVM::CallOp
2. Adds a pass that pattern matches the CallOp and lowers to LLVM code
