+++ include/triton/Conversion/TritonGPUToLLVM/TargetInfoBase.h
@@ -13,8 +13,8 @@ public:
   virtual Value ballot(ConversionPatternRewriter &rewriter, Location loc,
                        Type type, Value cmp) const = 0;
 
-  virtual Value storeShared(ConversionPatternRewriter &rewriter, Location loc,
-                            Value ptr, Value val, Value pred) const = 0;
+  virtual void storeShared(ConversionPatternRewriter &rewriter, Location loc,
+                           Value ptr, Value val, Value pred) const = 0;
   virtual Value loadShared(ConversionPatternRewriter &rewriter, Location loc,
                            const TypeConverter *converter, Value ptr,
                            Type elemTy, Value pred) const = 0;
